requirements.txt
streamlit==1.30.0
lightrag==0.1.5
openai==1.5.0
python-dotenv==1.0.0
asyncio==3.4.3 